+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Master
 
  * s/`function`/`kind`/ for `aggregators` in experiment files
+ * s/`aggregatorName`/`aggregator`/ for `sortOrder` in experiment files
  * Add `rules` to experiment files to allow user defined rules
 
 ## 0.1 (7th July 2017)